move the intervention policy card
