+++ packages/client/hmi-client/src/components/intervention-policy/tera-intervention-summary-card.vue
